chore(users): add mutation to request follow
When a user requests to follow another a UserConnection is inserted
in db. This row contains the user requesting the follow (userA) and
the user wanted to be followed (userB).

In a future commit a mutation will be implemented to accept the
request. In the above example, if userB accepts, UserConnection field
is_accepted will be set to True